Consolidate jagged_dense_bmm_bwd dispatch and perf tests into unified op_tests module
Merge the backward dispatch correctness test and the standalone Triton/FlyDSL perf bench into op_tests/test_jagged_dense_bmm_bwd.py,
and remove op_tests/flydsl_tests/test_jdbba_bwd_dispatch.py and op_tests/flydsl_tests/bench_jagged_dense_bmm_bwd_perf.py.

Correctness (pytest, Mi=512): Keeps dispatch config resolution, multi-D coverage (256 + 512 in one process), cos-similarity checks
on all three grads, the tail-over-read regression, forced split=2 reduce-path cases, and autograd (D≤256).

Performance (main(), Mi=7680): Replaces triton.testing.perf_report with the repo-standard @benchmark() + run_perftest flow.
Sweeps headline deployment shapes against FlyDSL and upstream Triton, with -c component (jagged / dense_bias / all) and -r regime
(default uniform; skew and genrec opt-in). Emits a markdown summary table with us, TFLOPS, TB/s, and err.

CLI: --worker-d runs the per-D correctness worker for manual debugging; default script invocation runs the perf sweep only (same
split as the forward jdbba test). Shared helpers (_make_inputs, torch reference, provider builders) live in the unified module.
